Merging new features from Develop

+++ server/src/controllers/diet-controller.ts
@@ -19,8 +19,7 @@ export const createDiet = async (req: Request, res: Response) => {
     if (!parsed.success) {
       return res.status(400).json({ message: "Invalid diet data", error: parsed.error });
     }
-    const { name, caloriesAmount } = parsed.data;
-    const newDiet = new DietModel({ name, caloriesAmount });
+    const newDiet = new DietModel(parsed.data);
     await newDiet.save();
     res.status(201).json(newDiet);
   } catch (error) {
@@ -48,6 +47,15 @@ export const updateDiet = async (req: Request, res: Response) => {
     if (!parsed.success) {
       return res.status(400).json({ message: "Invalid diet data", error: parsed.error });
     }
+
+    // retrieve user ID from the DB
+    const diet = await DietModel.findById(id);
+    if (!diet) return res.status(404).json({ message: "Diet not found" });
+
+    if (diet.userId.toString() !== parsed.data.userId) {
+      return res.status(403).json({ message: "Forbidden: not your diet" });
+    }
+
     const { name, caloriesAmount } = parsed.data;
     const updatedDiet = await DietModel.findByIdAndUpdate(
       id,
@@ -61,14 +69,34 @@ export const updateDiet = async (req: Request, res: Response) => {
   }
 };
 
-// DELETE /diets/:id - Delete a diet
+// DELETE /diets/:id/user/:userId - Delete a diet
 export const deleteDiet = async (req: Request, res: Response) => {
   try {
-    const { id } = req.params;
+    const { id, userId } = req.params;
+
+    const diet = await DietModel.findById(id);
+    if (!diet) return res.status(404).json({ message: "Diet not found" });
+
+    if (diet.userId.toString() !== userId) {
+      return res.status(403).json({ message: "Forbidden: not your diet" });
+    }
+
     const deleted = await DietModel.findByIdAndDelete(id);
     if (!deleted) return res.status(404).json({ message: "Diet not found" });
     res.sendStatus(204);
   } catch (error) {
     res.status(500).json({ message: "Error deleting diet", error });
+  }
+};
+
+// GET /diets/user/:userId - Get diets by User ID
+export const getDietsByUserId = async (req: Request, res: Response) => {
+  try {
+    const { userId } = req.params;
+    const diets = await DietModel.find({ userId });
+    res.status(200).json(diets);
+  } catch (error) {
+    console.log(error);
+    res.status(500).json({ message: "Error fetching diets for user", error });
   }
 };
